requirements.txt
absl-py==0.7.1
albumentations==0.4.5
antiorm==1.2.1
apex==0.1
appdirs==1.4.3
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.3.9
chardet==3.0.4
conda==4.3.16
cycler==0.10.0
Cython==0.29.12
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.0
defusedxml==0.6.0
device==0.1
dominate==2.4.0
entrypoints==0.3
fused==0.0.0
future==0.17.1
gast==0.2.2
geomloss==0.2.3
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
graphsurgeon==0.4.0
graphviz==0.13.2
grpcio==1.27.2
h5py==2.9.0
hypothesis==4.23.6
idna==2.8
imageio==2.5.0
imgaug==0.2.6
ipdb==0.13.2
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.1.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-property==0.0.1
lmdb==0.98
Mako==1.0.10
Markdown==3.1
MarkupSafe==1.1.1
-e git+https://github.com/facebookresearch/maskrcnn-benchmark.git@55796a04ea770029a80cf5933cc5c3f3f6fa59cf#egg=maskrcnn_benchmark
matplotlib==3.0.3
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.4
ninja==1.9.0.post1
notebook==5.7.8
numpy==1.16.3
oauthlib==3.1.0
opencv-python==4.1.1.26
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.11.0
POT==0.6.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycosat==0.6.3
Pygments==2.4.0
pykeops==1.2
pyparsing==2.4.0
pyrsistent==0.15.1
pysqlite3==0.4.2
pysqlite3-binary==0.4.2
pytest==4.5.0
python-dateutil==2.8.0
pytools==2019.1.1
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.15.93
scikit-image==0.16.2
scikit-learn==0.21.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
tensorboard==2.1.1
tensorboardX==1.6
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorrt==5.1.2.2
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tk==0.1.0
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.0.2
tqdm==4.32.2
traitlets==4.3.2
typing==3.6.6
uff==0.6.3
upfirdn2d==0.0.0
urllib3==1.24.2
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.12.1
yacs==0.1.6
